Move DiskWrite to filesystem module
