+++ filesystem/src/main/java/com/dropbox/android/external/fs3/Persister.kt
@@ -1,7 +1,5 @@
 package com.dropbox.android.external.fs3
 
-import com.dropbox.android.external.store4.DiskWrite
-
 /**
  * Interface for fetching data from persister
  *
