+++ store/api/store.api
@@ -1,7 +1,3 @@
-public abstract interface class com/dropbox/android/external/store4/DiskWrite {
-	public abstract fun write (Ljava/lang/Object;Ljava/lang/Object;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
 public abstract interface annotation class com/dropbox/android/external/store4/ExperimentalStoreApi : java/lang/annotation/Annotation {
 }
